+++ wp-content/plugins/core/src/Integrations/Gravity_Forms/Form_Markup.php
@@ -1,53 +1,26 @@
 <?php
+declare( strict_types=1 );
 
-namespace Tribe\Project\Theme;
+namespace Tribe\Project\Integrations\Gravity_Forms;
 
-/**
- * Class Forms
- *
- * Functions for handling forms, namely Gravity Forms
- *
- * @package Tribe\Project\Theme
- */
-class Gravity_Forms_Filter {
-
+class Form_Markup {
 	/**
 	 * @var bool Used to enable/disable CSS classes that control icon placement inside some field types.
 	 */
 	private $activate_icons = false;
 
 	/**
-	 * @return void
-	 * @action init
-	 */
-	public function hook() {
-		add_action( 'gform_enqueue_scripts', [ $this, 'enqueue_gravity_forms_jquery_ui_styles' ] );
-		add_filter( 'gform_field_choice_markup_pre_render', [ $this, 'customize_gf_choice_other' ], 10, 4 );
-		add_filter( 'gform_field_css_class', [ $this, 'add_gf_select_field_class' ], 10, 3 );
-		add_filter( 'gform_pre_render', [ $this, 'deactivate_gf_animations' ] );
-		add_filter( 'gform_confirmation_anchor', '__return_false' );
-		add_filter( 'gform_tabindex', '__return_false' );
-		add_filter( 'pre_option_rg_gforms_disable_css', '__return_true' );
-		add_filter( 'pre_option_rg_gforms_enable_html5', '__return_true' );
-	}
-
-	/**
-	 * Enqueue styles for datepicker on Gravity Forms
-	 */
-	public function enqueue_gravity_forms_jquery_ui_styles() {
-
-		global $wp_scripts;
-		$jquery_ui = $wp_scripts->query( 'jquery-ui-core' );
-		wp_enqueue_style( 'jquery-ui-smoothness',
-			'https://ajax.googleapis.com/ajax/libs/jqueryui/' . $jquery_ui->ver . '/themes/smoothness/jquery-ui.css',
-			false, 'screen' );
-
-	}
-
-	/**
 	 * Add some custom markup to other option for radio & checkbox controls
 	 *
-	 * @link https://www.gravityhelp.com/documentation/article/gform_field_choice_markup_pre_render/
+	 * @link   https://docs.gravityforms.com/gform_field_choice_markup_pre_render/
+	 * @filter gform_field_choice_markup_pre_render
+	 *
+	 * @param string $choice_markup
+	 * @param array  $choice
+	 * @param array  $field
+	 * @param string $value
+	 *
+	 * @return string
 	 */
 	public function customize_gf_choice_other( $choice_markup, $choice, $field, $value ) {
 
@@ -70,7 +43,14 @@ class Gravity_Forms_Filter {
 	/**
 	 * Add a custom class to the Gravity Forms select field
 	 *
-	 * @link http://www.gravityhelp.com/documentation/page/Gform_field_css_class
+	 * @link   https://docs.gravityforms.com/gform_field_css_class/
+	 * @filter gform_field_css_class
+	 *
+	 * @param string $classes
+	 * @param array  $field
+	 * @param array  $form
+	 *
+	 * @return string
 	 */
 	public function add_gf_select_field_class( $classes, $field, $form ) {
 
@@ -93,7 +73,7 @@ class Gravity_Forms_Filter {
 			$classes .= ' gf-textarea';
 		} elseif ( $field['type'] === 'date' || $field['inputType'] === 'date' ) {
 			$class_date_icon = ( $field['dateType'] === 'datepicker' ) ? $class_icon_simple : '';
-			$classes .= ' gf-date gf-date-layout-' . $field['dateType'] . $class_date_icon;
+			$classes         .= ' gf-date gf-date-layout-' . $field['dateType'] . $class_date_icon;
 		} elseif ( $field['type'] === 'time' || $field['inputType'] === 'time' ) {
 			$classes .= ' gf-time';
 		} elseif ( $field['type'] === 'phone' || $field['inputType'] === 'phone' ) {
@@ -115,24 +95,4 @@ class Gravity_Forms_Filter {
 		return $classes;
 
 	}
-
-	/**
-	 * Set enableAnimation for all forms to be false always. Removes a weird
-	 * collision between Gravity and GSAP jQuery plugin, our JS animation library, which is
-	 * worth the performance gain in animations used site wide over the Gravity
-	 * Forms conditional animations, which look terrible anyway.
-	 *
-	 * @param array $form
-	 *
-	 * @return array
-	 */
-	public function deactivate_gf_animations( $form = [] ) {
-
-		if ( isset( $form['enableAnimation'] ) && $form['enableAnimation'] == true ) {
-			$form['enableAnimation'] = false;
-		}
-
-		return $form;
-	}
-
 }
